created gettr and settr methods for adding species and reaction attributes to NapistuGraphs
