+++ src/napistu/network/constants.py
@@ -57,6 +57,16 @@ VALID_NAPISTU_GRAPH_NODE_TYPES = [
     NAPISTU_GRAPH_NODE_TYPES.SPECIES,
 ]
 
+NAPISTU_METADATA_KEYS = SimpleNamespace(
+    REACTION_ATTRS="reaction_attrs",
+    SPECIES_ATTRS="species_attrs",
+)
+
+ENTITIES_TO_ATTRS = {
+    SBML_DFS.REACTIONS: NAPISTU_METADATA_KEYS.REACTION_ATTRS,
+    SBML_DFS.SPECIES: NAPISTU_METADATA_KEYS.SPECIES_ATTRS,
+}
+
 # translating an SBML_dfs -> NapistuGraph
 
 GRAPH_WIRING_APPROACHES = SimpleNamespace(
