+++ src/napistu/network/net_create.py
@@ -9,10 +9,10 @@ import igraph as ig
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-from pydantic import BaseModel
 
 from napistu import sbml_dfs_core
 from napistu import utils
+from napistu.network import ng_utils
 from napistu.network import net_create_utils
 from napistu.network.ng_core import NapistuGraph
 
@@ -22,7 +22,6 @@ from napistu.constants import (
     SBO_MODIFIER_NAMES,
     SBOTERM_NAMES,
     SBML_DFS,
-    ENTITIES_W_DATA,
 )
 
 from napistu.network.constants import (
@@ -34,8 +33,6 @@ from napistu.network.constants import (
     NAPISTU_WEIGHTING_STRATEGIES,
     VALID_GRAPH_WIRING_APPROACHES,
     VALID_WEIGHTING_STRATEGIES,
-    DEFAULT_WT_TRANS,
-    DEFINED_WEIGHT_TRANSFORMATION,
     SCORE_CALIBRATION_POINTS_DICT,
     SOURCE_VARS_DICT,
     DROP_REACTIONS_WHEN,
@@ -105,7 +102,7 @@ def create_napistu_graph(
 
     # fail fast if reaction_graph_attrs is not properly formatted
     for k in reaction_graph_attrs.keys():
-        _validate_entity_attrs(
+        ng_utils._validate_entity_attrs(
             reaction_graph_attrs[k], custom_transformations=custom_transformations
         )
 
@@ -474,160 +471,6 @@ def create_napistu_graph_wiring(
     return decorated_all_reaction_edges_df
 
 
-def pluck_entity_data(
-    sbml_dfs: sbml_dfs_core.SBML_dfs,
-    graph_attrs: dict[str, dict],
-    data_type: str,
-    custom_transformations: Optional[dict[str, callable]] = None,
-) -> pd.DataFrame | None:
-    """
-    Pluck Entity Attributes from an sbml_dfs based on a set of tables and variables to look for.
-
-    Parameters
-    ----------
-    sbml_dfs : sbml_dfs_core.SBML_dfs
-        A mechanistic model.
-    graph_attrs : dict
-        A dictionary of species/reaction attributes to pull out. If the requested
-        data_type ("species" or "reactions") is not present as a key, or if the value
-        is an empty dict, this function will return None (no error).
-    data_type : str
-        "species" or "reactions" to pull out species_data or reactions_data.
-    custom_transformations : dict[str, callable], optional
-        A dictionary mapping transformation names to functions. If provided, these
-        will be checked before built-in transformations. Example:
-            custom_transformations = {"square": lambda x: x**2}
-
-    Returns
-    -------
-    pd.DataFrame or None
-        A table where all extracted attributes are merged based on a common index or None
-        if no attributes were extracted. If the requested data_type is not present in
-        graph_attrs, or if the attribute dict is empty, returns None. This is intended
-        to allow optional annotation blocks.
-
-    Raises
-    ------
-    ValueError
-        If data_type is not valid or if requested tables/variables are missing.
-    """
-
-    if data_type not in ENTITIES_W_DATA:
-        raise ValueError(
-            f'"data_type" was {data_type} and must be in {", ".join(ENTITIES_W_DATA)}'
-        )
-
-    if data_type not in graph_attrs.keys():
-        logger.info(
-            f'No {data_type} annotations provided in "graph_attrs"; returning None'
-        )
-        return None
-
-    entity_attrs = graph_attrs[data_type]
-    # validating dict
-    _validate_entity_attrs(entity_attrs, custom_transformations=custom_transformations)
-
-    if len(entity_attrs) == 0:
-        logger.info(
-            f'No attributes defined for "{data_type}" in graph_attrs; returning None'
-        )
-        return None
-
-    data_type_attr = data_type + "_data"
-    entity_data_tbls = getattr(sbml_dfs, data_type_attr)
-
-    data_list = list()
-    for k, v in entity_attrs.items():
-        # v["table"] is always present if entity_attrs is non-empty and validated
-        if v["table"] not in entity_data_tbls.keys():
-            raise ValueError(
-                f"{v['table']} was defined as a table in \"graph_attrs\" but "
-                f'it is not present in the "{data_type_attr}" of the sbml_dfs'
-            )
-
-        if v["variable"] not in entity_data_tbls[v["table"]].columns.tolist():
-            raise ValueError(
-                f"{v['variable']} was defined as a variable in \"graph_attrs\" but "
-                f"it is not present in the {v['table']} of the \"{data_type_attr}\" of "
-                "the sbml_dfs"
-            )
-
-        entity_series = entity_data_tbls[v["table"]][v["variable"]].rename(k)
-        trans_name = v.get("trans", DEFAULT_WT_TRANS)
-        # Look up transformation
-        if custom_transformations and trans_name in custom_transformations:
-            trans_fxn = custom_transformations[trans_name]
-        elif trans_name in DEFINED_WEIGHT_TRANSFORMATION:
-            trans_fxn = globals()[DEFINED_WEIGHT_TRANSFORMATION[trans_name]]
-        else:
-            # This should never be hit if _validate_entity_attrs is called correctly.
-            raise ValueError(
-                f"Transformation '{trans_name}' not found in custom_transformations or DEFINED_WEIGHT_TRANSFORMATION."
-            )
-        entity_series = entity_series.apply(trans_fxn)
-        data_list.append(entity_series)
-
-    if len(data_list) == 0:
-        return None
-
-    return pd.concat(data_list, axis=1)
-
-
-def apply_weight_transformations(
-    edges_df: pd.DataFrame, reaction_attrs: dict, custom_transformations: dict = None
-):
-    """
-    Apply Weight Transformations to edge attributes.
-
-    Parameters
-    ----------
-    edges_df : pd.DataFrame
-        A table of edges and their attributes extracted from a cpr_graph.
-    reaction_attrs : dict
-        A dictionary of attributes identifying weighting attributes within
-        an sbml_df's reaction_data, how they will be named in edges_df (the keys),
-        and how they should be transformed (the "trans" aliases).
-    custom_transformations : dict, optional
-        A dictionary mapping transformation names to functions. If provided, these
-        will be checked before built-in transformations.
-
-    Returns
-    -------
-    pd.DataFrame
-        edges_df with weight variables transformed.
-
-    Raises
-    ------
-    ValueError
-        If a weighting variable is missing or transformation is not found.
-    """
-
-    _validate_entity_attrs(
-        reaction_attrs, custom_transformations=custom_transformations
-    )
-
-    transformed_edges_df = copy.deepcopy(edges_df)
-    for k, v in reaction_attrs.items():
-        if k not in transformed_edges_df.columns:
-            raise ValueError(f"A weighting variable {k} was missing from edges_df")
-
-        trans_name = v["trans"]
-        # Look up transformation
-        if custom_transformations and trans_name in custom_transformations:
-            trans_fxn = custom_transformations[trans_name]
-        elif trans_name in DEFINED_WEIGHT_TRANSFORMATION:
-            trans_fxn = globals()[DEFINED_WEIGHT_TRANSFORMATION[trans_name]]
-        else:
-            # This should never be hit if _validate_entity_attrs is called correctly.
-            raise ValueError(
-                f"Transformation '{trans_name}' not found in custom_transformations or DEFINED_WEIGHT_TRANSFORMATION."
-            )
-
-        transformed_edges_df[k] = transformed_edges_df[k].apply(trans_fxn)
-
-    return transformed_edges_df
-
-
 def summarize_weight_calibration(
     napistu_graph: NapistuGraph, reaction_attrs: dict
 ) -> None:
@@ -647,7 +490,7 @@ def summarize_weight_calibration(
     """
 
     score_calibration_df = pd.DataFrame(SCORE_CALIBRATION_POINTS_DICT)
-    score_calibration_df_calibrated = apply_weight_transformations(
+    score_calibration_df_calibrated = ng_utils.apply_weight_transformations(
         score_calibration_df, reaction_attrs
     )
 
@@ -698,7 +541,7 @@ def add_graph_weights(
 
     napistu_graph_updated = copy.deepcopy(napistu_graph)
 
-    _validate_entity_attrs(reaction_attrs)
+    ng_utils._validate_entity_attrs(reaction_attrs)
 
     if weighting_strategy not in VALID_WEIGHTING_STRATEGIES:
         raise ValueError(
@@ -820,7 +663,7 @@ def _add_graph_weights_mixed(
 
     edges_df = napistu_graph.get_edge_dataframe()
 
-    calibrated_edges = apply_weight_transformations(edges_df, reaction_attrs)
+    calibrated_edges = ng_utils.apply_weight_transformations(edges_df, reaction_attrs)
     calibrated_edges = _create_source_weights(calibrated_edges, "source_wt")
 
     score_vars = list(reaction_attrs.keys())
@@ -869,7 +712,7 @@ def _add_graph_weights_calibration(
 
     edges_df = napistu_graph.get_edge_dataframe()
 
-    calibrated_edges = apply_weight_transformations(edges_df, reaction_attrs)
+    calibrated_edges = ng_utils.apply_weight_transformations(edges_df, reaction_attrs)
 
     score_vars = list(reaction_attrs.keys())
     score_vars.append("topo_weights")
@@ -1161,62 +1004,6 @@ def _create_source_weights(
     return source_wt_edges_df
 
 
-def _wt_transformation_identity(x):
-    """
-    Identity transformation for weights.
-
-    Parameters
-    ----------
-    x : any
-        Input value.
-
-    Returns
-    -------
-    any
-        The input value unchanged.
-    """
-    return x
-
-
-def _wt_transformation_string(x):
-    """
-    Map STRING scores to a similar scale as topology weights.
-
-    Parameters
-    ----------
-    x : float
-        STRING score.
-
-    Returns
-    -------
-    float
-        Transformed STRING score.
-    """
-    return 250000 / np.power(x, 1.7)
-
-
-def _wt_transformation_string_inv(x):
-    """
-    Map STRING scores so they work with source weights.
-
-    Parameters
-    ----------
-    x : float
-        STRING score.
-
-    Returns
-    -------
-    float
-        Inverse transformed STRING score.
-    """
-    # string scores are bounded on [0, 1000]
-    # and score/1000 is roughly a probability that
-    # there is a real interaction (physical, genetic, ...)
-    # reported string scores are currently on [150, 1000]
-    # so this transformation will map these onto {6.67, 1}
-    return 1 / (x / 1000)
-
-
 def _augment_network_nodes(
     network_nodes: pd.DataFrame,
     sbml_dfs: sbml_dfs_core.SBML_dfs,
@@ -1280,7 +1067,7 @@ def _augment_network_nodes(
     )
 
     # assign species_data related attributes to s_id
-    species_graph_data = pluck_entity_data(
+    species_graph_data = ng_utils.pluck_entity_data(
         sbml_dfs,
         species_graph_attrs,
         "species",
@@ -1372,7 +1159,7 @@ def _augment_network_edges(
     )
 
     # add other attributes based on reactions data
-    reaction_graph_data = pluck_entity_data(
+    reaction_graph_data = ng_utils.pluck_entity_data(
         sbml_dfs,
         reaction_graph_attrs,
         SBML_DFS.REACTIONS,
@@ -1639,58 +1426,3 @@ def _create_topology_weights(
         ]
 
     return napistu_graph
-
-
-def _validate_entity_attrs(
-    entity_attrs: dict,
-    validate_transformations: bool = True,
-    custom_transformations: Optional[dict] = None,
-) -> None:
-    """
-    Validate that graph attributes are a valid format.
-
-    Parameters
-    ----------
-    entity_attrs : dict
-        Dictionary of entity attributes to validate.
-    validate_transformations : bool, optional
-        Whether to validate transformation names, by default True.
-    custom_transformations : dict, optional
-        Dictionary of custom transformation functions, by default None. Keys are transformation names, values are transformation functions.
-
-    Returns
-    -------
-    None
-
-    Raises
-    ------
-    AssertionError
-        If entity_attrs is not a dictionary.
-    ValueError
-        If a transformation is not found in DEFINED_WEIGHT_TRANSFORMATION or custom_transformations.
-    """
-    assert isinstance(entity_attrs, dict), "entity_attrs must be a dictionary"
-
-    for k, v in entity_attrs.items():
-        # check structure against pydantic config
-        validated_attrs = _EntityAttrValidator(**v).model_dump()
-
-        if validate_transformations:
-            trans_name = validated_attrs.get("trans", DEFAULT_WT_TRANS)
-            valid_trans = set(DEFINED_WEIGHT_TRANSFORMATION.keys())
-            if custom_transformations:
-                valid_trans = valid_trans.union(set(custom_transformations.keys()))
-            if trans_name not in valid_trans:
-                raise ValueError(
-                    f"transformation '{trans_name}' was not defined as an alias in "
-                    "DEFINED_WEIGHT_TRANSFORMATION or custom_transformations. The defined transformations "
-                    f"are {', '.join(sorted(valid_trans))}"
-                )
-
-    return None
-
-
-class _EntityAttrValidator(BaseModel):
-    table: str
-    variable: str
-    trans: Optional[str] = DEFAULT_WT_TRANS
