+++ src/tests/test_network_net_create.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 import os
 
-import numpy as np
-import pandas as pd
 import pandas.testing as pdt
 import pytest
 
@@ -15,8 +13,6 @@ from napistu.network import ng_utils
 from napistu.constants import SBML_DFS
 from napistu.network.constants import (
     DROP_REACTIONS_WHEN,
-    DEFAULT_WT_TRANS,
-    WEIGHTING_SPEC,
     GRAPH_WIRING_APPROACHES,
     NAPISTU_GRAPH_EDGES,
     VALID_GRAPH_WIRING_APPROACHES,
@@ -211,140 +207,3 @@ def test_reverse_network_edges(reaction_species_examples):
     )
 
     assert net_create._reverse_network_edges(augmented_network_edges).shape[0] == 2
-
-
-def test_entity_validation():
-    # Test basic validation
-    entity_attrs = {"table": "reactions", "variable": "foo"}
-    assert net_create._EntityAttrValidator(**entity_attrs).model_dump() == {
-        **entity_attrs,
-        **{"trans": DEFAULT_WT_TRANS},
-    }
-
-    # Test validation with custom transformations
-    custom_transformations = {
-        "nlog10": lambda x: -np.log10(x),
-        "square": lambda x: x**2,
-    }
-
-    # Test valid custom transformation
-    entity_attrs_custom = {
-        "attr1": {
-            WEIGHTING_SPEC.TABLE: "reactions",
-            WEIGHTING_SPEC.VARIABLE: "foo",
-            WEIGHTING_SPEC.TRANSFORMATION: "nlog10",
-        },
-        "attr2": {
-            WEIGHTING_SPEC.TABLE: "species",
-            WEIGHTING_SPEC.VARIABLE: "bar",
-            WEIGHTING_SPEC.TRANSFORMATION: "square",
-        },
-    }
-    # Should not raise any errors
-    net_create._validate_entity_attrs(
-        entity_attrs_custom, custom_transformations=custom_transformations
-    )
-
-    # Test invalid transformation
-    entity_attrs_invalid = {
-        "attr1": {
-            WEIGHTING_SPEC.TABLE: "reactions",
-            WEIGHTING_SPEC.VARIABLE: "foo",
-            WEIGHTING_SPEC.TRANSFORMATION: "invalid_trans",
-        }
-    }
-    with pytest.raises(ValueError) as excinfo:
-        net_create._validate_entity_attrs(
-            entity_attrs_invalid, custom_transformations=custom_transformations
-        )
-    assert "transformation 'invalid_trans' was not defined" in str(excinfo.value)
-
-    # Test with validate_transformations=False
-    # Should not raise any errors even with invalid transformation
-    net_create._validate_entity_attrs(
-        entity_attrs_invalid, validate_transformations=False
-    )
-
-    # Test with non-dict input
-    with pytest.raises(AssertionError) as excinfo:
-        net_create._validate_entity_attrs(["not", "a", "dict"])
-    assert "entity_attrs must be a dictionary" in str(excinfo.value)
-
-
-def test_pluck_entity_data_species_identity(sbml_dfs):
-    # Take first 10 species IDs
-    species_ids = sbml_dfs.species.index[:10]
-    # Create mock data with explicit dtype to ensure cross-platform consistency
-    # Fix for issue-42: Use explicit dtypes to avoid platform-specific dtype differences
-    # between Windows (int32) and macOS/Linux (int64)
-    mock_df = pd.DataFrame(
-        {
-            "string_col": [f"str_{i}" for i in range(10)],
-            "mixed_col": np.arange(-5, 5, dtype=np.int64),  # Explicitly use int64
-            "ones_col": np.ones(10, dtype=np.float64),  # Explicitly use float64
-            "squared_col": np.arange(10, dtype=np.int64),  # Explicitly use int64
-        },
-        index=species_ids,
-    )
-    # Assign to species_data
-    sbml_dfs.species_data["mock_table"] = mock_df
-
-    # Custom transformation: square
-    def square(x):
-        return x**2
-
-    custom_transformations = {"square": square}
-    # Create graph_attrs for species
-    graph_attrs = {
-        "species": {
-            "string_col": {
-                WEIGHTING_SPEC.TABLE: "mock_table",
-                WEIGHTING_SPEC.VARIABLE: "string_col",
-                WEIGHTING_SPEC.TRANSFORMATION: "identity",
-            },
-            "mixed_col": {
-                WEIGHTING_SPEC.TABLE: "mock_table",
-                WEIGHTING_SPEC.VARIABLE: "mixed_col",
-                WEIGHTING_SPEC.TRANSFORMATION: "identity",
-            },
-            "ones_col": {
-                WEIGHTING_SPEC.TABLE: "mock_table",
-                WEIGHTING_SPEC.VARIABLE: "ones_col",
-                WEIGHTING_SPEC.TRANSFORMATION: "identity",
-            },
-            "squared_col": {
-                WEIGHTING_SPEC.TABLE: "mock_table",
-                WEIGHTING_SPEC.VARIABLE: "squared_col",
-                WEIGHTING_SPEC.TRANSFORMATION: "square",
-            },
-        }
-    }
-    # Call pluck_entity_data with custom transformation
-    result = net_create.pluck_entity_data(
-        sbml_dfs, graph_attrs, "species", custom_transformations=custom_transformations
-    )
-    # Check output
-    assert isinstance(result, pd.DataFrame)
-    assert set(result.columns) == {"string_col", "mixed_col", "ones_col", "squared_col"}
-    assert list(result.index) == list(species_ids)
-    # Check values
-    pd.testing.assert_series_equal(result["string_col"], mock_df["string_col"])
-    pd.testing.assert_series_equal(result["mixed_col"], mock_df["mixed_col"])
-    pd.testing.assert_series_equal(result["ones_col"], mock_df["ones_col"])
-    pd.testing.assert_series_equal(
-        result["squared_col"], mock_df["squared_col"].apply(square)
-    )
-
-
-def test_pluck_entity_data_missing_species_key(sbml_dfs):
-    # graph_attrs does not contain 'species' key
-    graph_attrs = {}
-    result = net_create.pluck_entity_data(sbml_dfs, graph_attrs, SBML_DFS.SPECIES)
-    assert result is None
-
-
-def test_pluck_entity_data_empty_species_dict(sbml_dfs):
-    # graph_attrs contains 'species' key but value is empty dict
-    graph_attrs = {SBML_DFS.SPECIES: {}}
-    result = net_create.pluck_entity_data(sbml_dfs, graph_attrs, SBML_DFS.SPECIES)
-    assert result is None
